Make rend2 build and work on linux

+++ codemp/rd-rend2/tr_bsp.cpp
@@ -30,6 +30,8 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 #include "tr_cache.h"
 #include <vector>
 
+#include <cmath>
+
 /*
 
 Loads and prepares a map file for scene rendering.
@@ -3575,7 +3577,7 @@ static std::vector<sprite_t> R_CreateSurfaceSpritesVertexData(
 		const vec2_t p01 = {p1[0] - p0[0], p1[1] - p0[1]};
 		const vec2_t p02 = {p2[0] - p0[0], p2[1] - p0[1]};
 
-		const float zarea = std::fabsf(p02[0]*p01[1] - p02[1]*p01[0]);
+		const float zarea = std::fabs(p02[0]*p01[1] - p02[1]*p01[0]);
 		if ( zarea <= 1.0 )
 		{
 			// Triangle's area is too small to consider.
@@ -3610,7 +3612,7 @@ static std::vector<sprite_t> R_CreateSurfaceSpritesVertexData(
 				// => y*y = 1.0 - x*x
 				// => y = -/+sqrt(1.0 - x*x)
 				float nx = flrand(-1.0f, 1.0f);
-				float ny = std::sqrtf(1.0f - nx*nx);
+				float ny = std::sqrt(1.0f - nx*nx);
 				ny *= irand(0, 1) ? -1 : 1;
 
 				VectorSet(sprite.normal, nx, ny, 0.0f);
